requirements.txt
# API and Protocol Dependencies
grpcio==1.54.0
protobuf==4.22.3
chirpstack-api==4.5.0

# Configuration Management
python-dotenv==1.0.0

# Logging and Date Handling
python-dateutil==2.8.2

# CLI Tools
click==8.1.3

# JSON Handling
jsonschema==4.17.3 